+++ lib/experimental/Utilities/Layout/exports.tsx
@@ -1,13 +1,7 @@
 import { Component } from "@/lib/component"
 import { AutoGrid as AutoGridComponent } from "./AutoGrid"
-import {
-  InfoPaneLayout as InfoPaneLayoutComponent,
-  InfoPaneLayoutProps,
-} from "./InfoPaneLayout"
 import { Split as SplitComponent } from "./Split"
 import { Stack as StackComponent } from "./Stack"
-
-export type { InfoPaneLayoutProps }
 
 export const Stack = Component(
   {
@@ -32,12 +26,4 @@ export const AutoGrid = Component(
   },
 
   AutoGridComponent
-)
-
-export const InfoPaneLayout = Component(
-  {
-    name: "InfoPaneLayout",
-    type: "layout",
-  },
-  InfoPaneLayoutComponent
 )
